RAG on the edge as websocket
