+++ agora/cerebral_api/src/websocket-client/config.js
@@ -1,0 +1,9 @@
+module.exports = {
+    SERVER_URL: 'http://localhost:5030',
+    INDUSTRIES: {
+        'default': ['default'],
+        'manufacturing': ['maintenance engineer', 'shift supervisor'],
+        'retail': ['store manager', 'buyer'],
+        'hypermarket': ['store manager', 'shopper', 'maintenance worker']
+    }
+};
