+++ agora/cerebral_api/src/websocket-client/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "websocket-client",
+  "version": "1.0.0",
+  "main": "config.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "readline": "^1.3.0",
+    "socket.io-client": "^4.8.1"
+  },
+  "description": ""
+}
